% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{as_kg_igraph}
\alias{as_kg_igraph}
\alias{as_kg_igraph.data.frame}
\alias{as_kg_igraph.tweetgraph_primitive}
\title{Build a knowledge \code{<igraph>} from a \code{<tweetgraph_primitive>} or tweet
\code{<data.frame>}.}
\usage{
as_kg_igraph(x, ...)

\method{as_kg_igraph}{data.frame}(x, ...)

\method{as_kg_igraph}{tweetgraph_primitive}(x, ...)
}
\arguments{
\item{x}{\code{<tweetgraph_primitive>} or tweet \code{<data.frame>}}

\item{...}{Arguments passed to or from other methods.}
}
\value{
\code{<igraph>}
}
\description{
Build a knowledge \code{<igraph>} from a \code{<tweetgraph_primitive>} or tweet
\code{<data.frame>}.
}
\examples{
\dontrun{


hashtag_rstats <- rtweet::search_tweets("#rstats")

as_kg_igraph(hashtag_rstats)


}

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
