% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge-graph.R
\name{as_kg_primitive}
\alias{as_kg_primitive}
\title{Build a knowledge graph \code{<tweetgraph_primitive>}.}
\usage{
as_kg_primitive(tweet_df, edge_direction = c("info_flow", "interaction"))
}
\arguments{
\item{tweet_df}{\itemize{
\item \code{data.frame}
\itemize{
\item as returned by  \href{https://rtweet.info/}{{rtweet}}
}
}}

\item{edge_direction}{\itemize{
\item \code{character(1L)}
\itemize{
\item \code{"info_flow"} or \code{"interaction"}
}
}}
}
\value{
\code{<tweetgraph_primitive>}
}
\description{
Build a knowledge graph \code{<tweetgraph_primitive>}.
}
\examples{
\dontrun{


hashtag_rstats <- rtweet::search_tweets("#rstats")

as_kg_primitive(hashtag_rstats)


}

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
