% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge-graph.R
\name{as_knowledge_graph}
\alias{as_knowledge_graph}
\title{Build a knowledge graph from a tweet data frame.}
\usage{
as_knowledge_graph(tweet_df, .edge_type = c("info_flow", "interaction"))
}
\arguments{
\item{tweet_df}{\itemize{
\item \code{data.frame}
\itemize{
\item as returned by  \href{https://rtweet.info/}{{rtweet}}
}
}}

\item{.edge_type}{\itemize{
\item \code{character(1L)}
\itemize{
\item \code{"info_flow"} or \code{"interaction"}
}
}}
}
\value{
\code{igraph}
}
\description{
Build a knowledge graph from a tweet data frame.
}
\examples{
\dontrun{


hashtag_rstats <- rtweet::search_tweets("#rstats")

tweet_graph <- as_knowledge_graph(hashtag_rstats)


}

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
