% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{as_sna_igraph}
\alias{as_sna_igraph}
\alias{as_sna_igraph.data.frame}
\alias{as_sna_igraph.tweetgraph_primitive}
\title{Build an \code{<igraph>} for social network analysis from a
\code{<tweetgraph_primitive>} or tweet \code{<data.frame>}.}
\usage{
as_sna_igraph(x, ...)

\method{as_sna_igraph}{data.frame}(x, ...)

\method{as_sna_igraph}{tweetgraph_primitive}(x, ...)
}
\arguments{
\item{x}{\code{<tweetgraph_primitive>} or tweet \code{<data.frame>}}

\item{...}{Arguments passed to or from other methods.}
}
\value{
\code{<igraph>}
}
\description{
Build an \code{<igraph>} for social network analysis from a
\code{<tweetgraph_primitive>} or tweet \code{<data.frame>}.
}
\examples{
\dontrun{


hashtag_rstats <- rtweet::search_tweets("#rstats")

as_sna_igraph(hashtag_rstats)


}

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
