% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social-network.R
\name{as_social_network}
\alias{as_social_network}
\title{Build a graph ready for proper social network analysis.}
\usage{
as_social_network(tweet_df)
}
\arguments{
\item{tweet_df}{\itemize{
\item \code{data.frame}
\itemize{
\item as returned by  \href{https://rtweet.info/}{{rtweet}}
}
}}
}
\value{
\code{igraph}
}
\description{
Build a graph ready for proper social network analysis.
}
\examples{
\dontrun{


hashtag_rstats <- rtweet::search_tweets("#rstats")

tweet_graph <- as_social_network(hashtag_rstats)


}

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
