% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{as_socnet_igraph}
\alias{as_socnet_igraph}
\alias{as_socnet_igraph.data.frame}
\alias{as_socnet_igraph.tweetgraph_primitive}
\title{Build an \code{<igraph>} for social network analysis from a
\code{<tweetgraph_primitive>} or tweet \code{<data.frame>}.}
\usage{
as_socnet_igraph(x, ...)

\method{as_socnet_igraph}{data.frame}(x, action = c("all", "mentions",
  "retweet", "quoted", "reply_to"))

\method{as_socnet_igraph}{tweetgraph_primitive}(x, ...)
}
\arguments{
\item{x}{\code{<tweetgraph_primitive>} or tweet \code{<data.frame>}}

\item{action}{\code{<character>}
\itemize{
\item \code{"all"}, \code{"mentions"}, \code{"retweet"}, \code{"quoted"}, and/or \code{"reply_to"}
}}
}
\value{
\code{<igraph>}
}
\description{
Build an \code{<igraph>} for social network analysis from a
\code{<tweetgraph_primitive>} or tweet \code{<data.frame>}.
}
\examples{
\dontrun{


hashtag_rstats <- rtweet::search_tweets("#rstats")

as_socnet_igraph(hashtag_rstats)


}

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
