% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social-network.R
\name{as_socnet_primitive}
\alias{as_socnet_primitive}
\title{Build a social network \code{<tweetgraph_primitive>}.}
\usage{
as_socnet_primitive(tweet_df, action = c("all", "mentions", "retweet",
  "quoted", "reply_to"))
}
\arguments{
\item{tweet_df}{\itemize{
\item \code{data.frame}
\itemize{
\item as returned by  \href{https://rtweet.info/}{{rtweet}}
}
}}

\item{action}{\code{<character>}
\itemize{
\item \code{"all"}, \code{"mentions"}, \code{"retweet"}, \code{"quoted"}, and/or \code{"reply_to"}
}}
}
\value{
\code{<tweetgraph_primitive>}
}
\description{
Build a social network \code{<tweetgraph_primitive>}.
}
\examples{
\dontrun{


hashtag_rstats <- rtweet::search_tweets("#rstats")

as_social_network_primitive(hashtag_rstats)


}

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
