% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNetwork.R
\name{plot_vis_net}
\alias{plot_vis_net}
\title{Interactively visualize a graph of twitter data}
\usage{
plot_vis_net(tweet_graph)
}
\arguments{
\item{tweet_graph}{\itemize{
\item Currently objects obtained via \code{as_knowledge_graph()} and
\code{as_social_network()}
}}
}
\value{
\code{igraph}
}
\description{
Interactively visualize a graph of twitter data
}
\examples{
\dontrun{


hashtag_rstats <- rtweet::search_tweets("#rstats")

hashtag_rstats \%>\% 
  as_knowledge_graph() \%>\% 
  plot_vis_net()

hashtag_rstats \%>\% 
  as_social_network() \%>\% 
  plot_vis_net()

}

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
