% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social-network.R
\name{as_social_network}
\alias{as_social_network}
\title{Build a graph ready for social network analysis.}
\usage{
as_social_network(tweet_df, action = c("all", "mentions", "retweet",
  "quoted", "reply_to"))
}
\arguments{
\item{tweet_df}{\itemize{
\item \code{data.frame}
\itemize{
\item as returned by  \href{https://rtweet.info/}{{rtweet}}
}
}}

\item{action}{\itemize{
\item \code{character}
\item \code{"all"}, \code{"mentions"}, \code{"retweet"}, \code{"quoted"}, and/or \code{"reply_to"}
}}
}
\value{
\code{igraph}
}
\description{
Build a graph ready for social network analysis.
}
\examples{
\dontrun{


hashtag_rstats <- rtweet::search_tweets("#rstats")

tweet_graph <- as_social_network(hashtag_rstats)


}

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
